Renamed AddItemInputWidget to AddIngredientWidget
